backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
pydantic==2.8.2
pydantic-settings==2.4.0
sqlmodel==0.0.21
SQLAlchemy==2.0.30
httpx==0.27.0
python-multipart==0.0.9
pypdf==4.0.1
python-docx==1.1.0
python-pptx==0.6.23
beautifulsoup4==4.12.3
lxml==5.1.0
chromadb==0.4.22
sentence-transformers==2.3.1
tiktoken==0.6.0
langchain-text-splitters==0.0.1
rank-bm25==0.2.2
numpy==1.26.4
openai==1.40.6
cohere==4.47
google-generativeai==0.8.3
ollama==0.3.2
requests==2.31.0
crawl4ai==0.2.77
python-dotenv==1.0.1
